requirements.txt
Flask
newspaper3k
deep-translator
lxml